Add files via upload
Environment C is one of the scenarios in the base paper we are working. This version of code has considered the Environment C obstacles are implemented Bi-RRT star algorithm in that scenario with an additional bonus boundary. While plotting the path, the bonus boundary needs to be taken into account to avoid almost collision.